requirements.txt
sklearn
torch>=1.7
torchvision
git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup
numpy
pillow
matplotlib
timm
scikit-image
grad-cam
opencv-python
ttach